Extract `crates_io_diesel_helpers` crate
This makes it possible to compile them concurrently and also makes them potentially usable in other workspace crates.
