+++ crates/crates_io_diesel_helpers/src/pg_enum.rs
@@ -1,30 +1,11 @@
-use diesel::sql_types::{Date, Double, Integer, Interval, SingleValue, Text, Timestamp};
-
-mod semver;
-
-pub use semver::SemverVersion;
-
-define_sql_function!(#[aggregate] fn array_agg<T: SingleValue>(x: T) -> Array<T>);
-define_sql_function!(fn canon_crate_name(x: Text) -> Text);
-define_sql_function!(fn to_char(a: Date, b: Text) -> Text);
-define_sql_function!(fn lower(x: Text) -> Text);
-define_sql_function!(fn date_part(x: Text, y: Timestamp) -> Double);
-define_sql_function! {
-    #[sql_name = "date_part"]
-    fn interval_part(x: Text, y: Interval) -> Double;
-}
-define_sql_function!(fn floor(x: Double) -> Integer);
-define_sql_function!(fn greatest<T: SingleValue>(x: T, y: T) -> T);
-define_sql_function!(fn least<T: SingleValue>(x: T, y: T) -> T);
-define_sql_function!(fn split_part(string: Text, delimiter: Text, n: Integer) -> Text);
-
+#[macro_export]
 macro_rules! pg_enum {
     (
         $vis:vis enum $name:ident {
             $($item:ident = $int:expr,)*
         }
     ) => {
-        #[derive(Debug, Copy, Clone, PartialEq, Eq, Hash, Serialize, Deserialize, FromSqlRow, AsExpression)]
+        #[derive(Debug, Copy, Clone, PartialEq, Eq, Hash, serde::Serialize, serde::Deserialize, diesel::FromSqlRow, diesel::AsExpression)]
         #[diesel(sql_type = diesel::sql_types::Integer)]
         #[serde(rename_all = "snake_case")]
         #[repr(i32)]
@@ -55,5 +36,3 @@ macro_rules! pg_enum {
         }
     }
 }
-
-pub(crate) use pg_enum;
