+++ src/controllers/krate/downloads.rs
@@ -7,11 +7,11 @@ use crate::app::AppState;
 use crate::controllers::krate::CratePath;
 use crate::models::{Version, VersionDownload};
 use crate::schema::{version_downloads, versions};
-use crate::sql::to_char;
 use crate::util::errors::AppResult;
 use crate::views::EncodableVersionDownload;
 use axum_extra::json;
 use axum_extra::response::ErasedJson;
+use crates_io_diesel_helpers::to_char;
 use diesel::prelude::*;
 use diesel_async::RunQueryDsl;
 use std::cmp;
