+++ src/controllers/user/other.rs
@@ -8,9 +8,9 @@ use diesel_async::RunQueryDsl;
 use crate::app::AppState;
 use crate::models::{CrateOwner, OwnerKind, User};
 use crate::schema::{crate_downloads, crate_owners, crates};
-use crate::sql::lower;
 use crate::util::errors::AppResult;
 use crate::views::EncodablePublicUser;
+use crates_io_diesel_helpers::lower;
 
 /// Find user by login.
 #[utoipa::path(
