+++ src/rate_limiter.rs
@@ -1,7 +1,7 @@
 use crate::schema::{publish_limit_buckets, publish_rate_overrides};
-use crate::sql::{date_part, floor, greatest, interval_part, least, pg_enum};
 use crate::util::errors::{AppResult, TooManyRequests};
 use chrono::{NaiveDateTime, Utc};
+use crates_io_diesel_helpers::{date_part, floor, greatest, interval_part, least, pg_enum};
 use diesel::dsl::IntervalDsl;
 use diesel::prelude::*;
 use diesel::sql_types::Interval;
